Moved dtos to different folder
