+++ services/hype-backend/src/modules/reward/dto/index.ts
@@ -1,0 +1,4 @@
+export * from './claim.dto';
+export * from './impression.dto';
+export * from './reward.dto';
+export * from './rewardState.dto';
