+++ services/hype-backend/src/modules/reward/dto/reward.dto.ts
