+++ services/hype-backend/src/modules/reward/pools.controller.ts
@@ -22,8 +22,7 @@ import { catchError, firstValueFrom, map } from 'rxjs';
 import { AxiosResponse } from '@nestjs/terminus/dist/health-indicator/http/axios.interfaces';
 import * as dotenv from 'dotenv';
 import { WalletGuard, HmacMiddleware } from '../guards';
-import { ImpressionDto } from './impression.dto';
-import { RewardDto } from './reward.dto';
+import { ImpressionDto, RewardDto } from './dto';
 import { RewardService } from './reward.service';
 import impressionsJson from '../../../pool_impressions_example.json';
 import { IPool } from '../../models';
